New op Array2
Made to replace the old array op which fills an array with one value
and the ContinuousNumberArray op which filled an array with incrementing
values.
op has a dropdown box to select the mode.
When 0,1,2,3 mode is selected then the default value paramater is greyed
out.
Example : https://cables.gl/ui/#/project/2cL8wH

If this op is merged then the old array and ContinuousNumberArray
should be deprecated ticket #791